Gen-ai and llm/Chatbot-Execute4.0/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-core
langchain-community==0.3.19
transformers
torch
huggingface_hub
langchain_huggingface

